refactor: multi-fed support and code clean up
