+++ apps/gateway-ui/src/App.tsx
@@ -31,7 +31,6 @@ export const App = React.memo(function Admin(): JSX.Element {
     route_hints: [],
     version_hash: '',
   });
-  const [federationId, setFederationId] = useState<string | null>(null);
   const [authenticated, setAuthenticated] = useState<boolean>(false);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState<string>('');
@@ -59,11 +58,6 @@ export const App = React.memo(function Admin(): JSX.Element {
           .fetchInfo()
           .then((gatewayInfo) => {
             setGatewayInfo(gatewayInfo);
-            setFederationId(
-              federationId
-                ? federationId
-                : gatewayInfo.federations[0].federation_id
-            );
           })
           .catch(({ message, error }) => {
             console.error(error);
@@ -76,7 +70,7 @@ export const App = React.memo(function Admin(): JSX.Element {
       const interval = setInterval(fetchInfo, 5000);
       return () => clearInterval(interval);
     }
-  }, [gateway, authenticated, federationId]);
+  }, [gateway, authenticated]);
 
   const content = useMemo(() => {
     if (loading) {
@@ -131,46 +125,29 @@ export const App = React.memo(function Admin(): JSX.Element {
 
     return (
       <Box>
-        {gatewayInfo.federations.length === 0 && (
-          <>
-            <ConnectFederation
-              renderConnectedFedCallback={(federation: Federation) => {
-                setGatewayInfo({
-                  ...gatewayInfo,
-                  federations: [...gatewayInfo.federations, federation],
-                });
-              }}
-            />
-          </>
-        )}
-        <Flex direction='column' gap={8}>
-          {federationId && (
-            <FederationCard
-              key={federationId}
-              federation={
-                gatewayInfo.federations.find(
-                  (fed) => fed.federation_id === federationId
-                ) as Federation
-              }
-              setFederationId={setFederationId}
-              federations={gatewayInfo.federations}
-              network={gatewayInfo.network}
-              lightning_pub_key={gatewayInfo.lightning_pub_key}
-            />
-          )}
+        <ConnectFederation
+          renderConnectedFedCallback={(federation: Federation) => {
+            setGatewayInfo({
+              ...gatewayInfo,
+              federations: [...gatewayInfo.federations, federation],
+            });
+          }}
+        />
+        <Flex flexDirection={'column'} gap={8}>
+          {gatewayInfo.federations.map((federation: Federation) => {
+            return (
+              <FederationCard
+                key={federation.federation_id}
+                federation={federation}
+                network={gatewayInfo.network}
+                lightning_pub_key={gatewayInfo.lightning_pub_key}
+              />
+            );
+          })}
         </Flex>
       </Box>
     );
-  }, [
-    gateway,
-    loading,
-    authenticated,
-    error,
-    gatewayInfo,
-    theme,
-    t,
-    federationId,
-  ]);
+  }, [gateway, loading, authenticated, error, gatewayInfo, theme, t]);
 
   return (
     <ApiProvider props={{ gateway }}>
